+++ source/core/scene/material/glass/glass_thin_material.hpp
@@ -12,7 +12,7 @@ class Glass_thin : public Material {
                                    sampler::Sampler& sampler, Worker const& worker,
                                    uint32_t depth) const noexcept override final;
 
-    float3 thin_absorption(float3 const& wo, float3 const& n, float2 uv, float time,
+    float3 thin_absorption(float3 const& wo, float3 const& n, float2 uv, uint64_t time,
                            Sampler_filter filter, Worker const& worker) const
         noexcept override final;
